+++ index.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { Provider } from 'react-redux'
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import App from './src/containers/app';
 import configureStore from './src/store/index';
-import Views from './src/views/app';
 // import rootSaga from './sagas/index'
 
 let store = configureStore({});
@@ -11,7 +11,7 @@ let store = configureStore({});
 ReactDOM.render(
     <div className="container">
         <Provider store={store}>
-            <Views />
+            <App />
         </Provider>
     </div>,
     document.getElementById('app'));
