--- src/views/grid.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-export default React.createClass({
-    displayName: 'Grid',
-
-    render() {
-        return (
-            <div>
-                About
-            </div>
-        );
-    }
-});
